requirement.txt
agate==1.9.1
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
alembic==1.13.1
Babel==2.14.0
blinker==1.7.0
cachelib==0.12.0
click==8.1.7
colorama==0.4.6
csvkit==1.5.0
dbfread==2.0.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
leather==0.4.0
Mako==1.3.2
MarkupSafe==2.1.5
msgspec==0.18.6
mysql-connector-python==8.3.0
olefile==0.47
openpyxl==3.1.2
parsedatetime==2.6
PyMySQL==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
six==1.16.0
SQLAlchemy==2.0.29
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2
xlrd==2.0.1
